fixed v080 upgrade crash
v073 saved scripts in obfuscated form due to an error in my Proguard
configuration, so this is a fix to upgrade users from the obfuscated
models to the current models.

Since the database just saves JSON, I only had to update the models
themselves and resave the json right after converting.